Create email address object on account creation
When an account is created we need to make sure an email address is created
if asked (default) and User object has one defined. This fixes cases where
accounts are created with email authentication as those accounts are unable
to be logged in to.

Fixes #34 — thanks Luke Hatcher for the report.