Changed Account.timezone default and added Account.localtime
Account.timezone defaults to an empty string. This allows empty to represent
NOT SET and a site developer can use this to ask for user to set their account
timezone. An empty is handled with settings.TIME_ZONE as the behavior
originally did prior to this change.

Account learned how to convert a given datetime object into the timezone of
an account with Account.localtime.